fix(refactor): refactor code for isolation
