+++ components/cart/add-to-cart.tsx
@@ -4,7 +4,8 @@ import { PlusIcon } from '@heroicons/react/24/outline';
 import clsx from 'clsx';
 import { addItem } from 'components/cart/actions';
 import LoadingDots from 'components/loading-dots';
-import { ProductVariant, Store } from 'lib/shopify/types';
+import { Store } from 'lib/aspire/types';
+import { ProductVariant } from 'lib/shopify/types';
 import { useSearchParams } from 'next/navigation';
 import { useFormState, useFormStatus } from 'react-dom';
