+++ components/cart/delete-item-button.tsx
@@ -4,7 +4,8 @@ import { XMarkIcon } from '@heroicons/react/24/outline';
 import clsx from 'clsx';
 import { removeItem } from 'components/cart/actions';
 import LoadingDots from 'components/loading-dots';
-import type { CartItem, Store } from 'lib/shopify/types';
+import { Store } from 'lib/aspire/types';
+import type { CartItem } from 'lib/shopify/types';
 import { useFormState, useFormStatus } from 'react-dom';
 
 function SubmitButton() {
