+++ components/product/product-description.tsx
@@ -1,7 +1,8 @@
 import { AddToCart } from 'components/cart/add-to-cart';
 import Price from 'components/price';
 import Prose from 'components/prose';
-import { Product, Store } from 'lib/shopify/types';
+import { Store } from 'lib/aspire/types';
+import { Product } from 'lib/shopify/types';
 import { Suspense } from 'react';
 import { VariantSelector } from './variant-selector';
